requirements.txt
Flask==2.3.2
google-cloud-firestore==2.11.1
google-cloud-language==2.10.0
gunicorn==21.2.0